PDBUtility: Refactor this to a more central place
